Put delete button in edit page
